requirements.txt
google-adk[database]==1.17.0
psutil==7.1.2
google-generativeai==0.8.5
python-dotenv==1.2.1
beautifulsoup4==4.14.2
requests==2.32.5
sqlalchemy==2.0.44
